add possibility on UI to add and remove a profile

+++ backend/api/src/main/kotlin/com/akurat/model/ProfileResponse.kt
@@ -3,6 +3,6 @@ package com.akurat.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProfileResponse(val id: Long, val name: String, val role: String, val createdAt: Long)
+data class ProfileResponse(val id: Int, val name: String, val role: String, val createdAt: Long)
 
 fun Profile.toResponse() = ProfileResponse(id, name, role.name, createdAt.toEpochMilli())
