+++ backend/domain/src/main/kotlin/com/akurat/model/Profile.kt
@@ -2,4 +2,4 @@ package com.akurat.model
 
 import java.time.Instant
 
-data class Profile(val id: Long, val name: String, val role: Role, val createdAt: Instant)
+data class Profile(val id: Int, val name: String, val role: Role, val createdAt: Instant)
